+++ src/main/scala/representation/Chord.scala
@@ -5,12 +5,12 @@ object Chord {
 }
 
 case class Chord(notes: List[Note]) extends MusicalElement {
-  override def getDuration: Double = notes.head.duration
-  override def getStartTime: Double = notes.head.startTime
+  override def getDuration: BigDecimal = notes.head.duration
+  override def getStartTime: BigDecimal = notes.head.startTime
 
-  override def withDuration(duration: Double): MusicalElement =
+  override def withDuration(duration: BigDecimal): MusicalElement =
     copy(notes = notes.map(_.withDuration(duration)))
 
-  override def withStartTime(startTime: Double): MusicalElement =
+  override def withStartTime(startTime: BigDecimal): MusicalElement =
     copy(notes = notes.map(_.withStartTime(startTime)))
 }
