+++ src/main/scala/representation/Phrase.scala
@@ -10,7 +10,7 @@ object Phrase {
   val DEFAULT_TEMPO_BPM = 120
   val DEFAULT_START_TIME = 0.0
 
-  def computeDuration(phrase: Phrase): Double = phrase.sumBy(0.0, _.getDuration)
+  def computeDuration(phrase: Phrase): BigDecimal = phrase.sumBy(0.0, _.getDuration)
 
   def apply(): Phrase = {
     new Phrase()
@@ -35,10 +35,10 @@ case class Phrase(
   musicalElements: List[MusicalElement] = List(),
   polyphony: Boolean = false,
   tempoBPM: Double = Phrase.DEFAULT_TEMPO_BPM,
-  startTime: Double = Phrase.DEFAULT_START_TIME)
+  startTime: BigDecimal = Phrase.DEFAULT_START_TIME)
     extends MusicalElement with Traversable[MusicalElement] {
   private val maxChordSize: MemoizedFunc[Phrase, Int] = FunctionalUtils.memoized(Phrase.computeMaxChordSize)
-  private val duration: MemoizedFunc[Phrase, Double] = FunctionalUtils.memoized(Phrase.computeDuration)
+  private val duration: MemoizedFunc[Phrase, BigDecimal] = FunctionalUtils.memoized(Phrase.computeDuration)
 
   require(! polyphony || canHavePolyphony)
 
@@ -54,22 +54,22 @@ case class Phrase(
   def withMusicalElements(musicalElements: MusicalElement*) =
     copy(musicalElements = musicalElements.toList)
 
-  override def withDuration(newDuration: Double): MusicalElement = scaled(newDuration / getDuration)
+  override def withDuration(newDuration: BigDecimal): MusicalElement = scaled(newDuration / getDuration)
 
-  override def withStartTime(startTime: Double): MusicalElement = {
+  override def withStartTime(startTime: BigDecimal): MusicalElement = {
     val shiftedMusicalElements = musicalElements.map(_.withStartTime(startTime))
     copy(startTime = startTime, musicalElements = shiftedMusicalElements)
   }
 
   def getMaxChordSize: Int = maxChordSize(this)
 
-  def scaled(durationRatio: Double): Phrase = {
+  def scaled(durationRatio: BigDecimal): Phrase = {
     val scaledElems = musicalElements.map(elem => elem.withDuration(elem.getDuration * durationRatio))
     copy(musicalElements = scaledElems)
   }
 
-  override def getDuration: Double = duration(this)
+  override def getDuration: BigDecimal = duration(this)
   override def isEmpty: Boolean = musicalElements.isEmpty
   override def foreach[U](f: (MusicalElement) => U): Unit = musicalElements.foreach(f)
-  override def getStartTime: Double = startTime
+  override def getStartTime: BigDecimal = startTime
 }
