+++ src/main/scala/representation/Rest.scala
@@ -6,12 +6,12 @@ object Rest {
 }
 
 case class Rest(
-    duration: Double = 0.0,
-    startTime: Double = 0.0) extends MusicalElement {
-  override def getDuration: Double = duration
-  override def getStartTime: Double = startTime
+    duration: BigDecimal = 0.0,
+    startTime: BigDecimal = 0.0) extends MusicalElement {
+  override def getDuration: BigDecimal = duration
+  override def getStartTime: BigDecimal = startTime
 
-  override def withDuration(duration: Double): MusicalElement = copy(duration = duration)
+  override def withDuration(duration: BigDecimal): MusicalElement = copy(duration = duration)
 
-  override def withStartTime(startTime: Double): MusicalElement = copy(startTime = startTime)
+  override def withStartTime(startTime: BigDecimal): MusicalElement = copy(startTime = startTime)
 }
